既にある swf,jpg,png ファイルを上書きしないように修正 (無駄なので)
git-svn-id: svn+ssh://svn.sourceforge.jp/svnroot/swfed/trunk@84 7c90b180-03d5-4157-b861-58a559ae9d1e

+++ www/swfdump.php
@@ -22,8 +22,9 @@ if (! empty($_FILES['swffile']['tmp_name'])) {
     $tmp_name = sha1($swfdata, false);
     $id = substr($tmp_name, 0, 16); // XXX
     $tmp_filename = "$tmp_prefix$id.swf";
-    if (! file_put_contents($tmp_filename, $swfdata)) {
-        echo 'ERROR: file_put_contents failed.  zero size ?'."\n";
+    if ((! is_readable($tmp_filename)) &&
+        (! file_put_contents($tmp_filename, $swfdata))) {
+        fprintf(stderr, "swfdump.php: file_put_contents failed. zero size?\n");
         unlink($tmp_filename);
         exit(0);
     }
